+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/veryl/compare/v0.2.1...Unreleased) - ReleaseDate
 
+* [Added] attribute arguments check [#90](https://github.com/dalance/veryl/issues/90)
+
 ## [v0.2.1](https://github.com/dalance/veryl/compare/v0.2.0...v0.2.1) - 2023-01-19
 
 * [Added] uknown member check [#84](https://github.com/dalance/veryl/issues/84)